Developer reference

Developer reference

Internal global variables

Configuration-related variables

These are set during execution of Revise's __init__ function.

Revise.watching_files — Constant.

Revise.watching_files[]

Returns true if we watch files rather than their containing directory. FreeBSD and NFS-mounted systems should watch files, otherwise we prefer to watch directories.

Revise.polling_files — Constant.

Revise.polling_files[]

Returns true if we should poll the filesystem for changes to the files that define loaded code. It is preferable to avoid polling, instead relying on operating system notifications via FileWatching.watch_file. However, NFS-mounted filesystems (and perhaps others) do not support file-watching, so for code stored on such filesystems you should turn polling on.

See the documentation for the JULIA_REVISE_POLL environment variable.

Revise.tracking_Main_includes — Constant.

Revise.tracking_Main_includes[]

Returns true if files directly included from the REPL should be tracked. The default is false. See the documentation regarding the JULIA_REVISE_INCLUDE environment variable to customize it.

Path-related variables

Revise.juliadir — Constant.

Revise.juliadir

Constant specifying full path to julia top-level directory from which julia was built. This is reliable even for cross-builds.

Revise.basesrccache — Constant.

Revise.basesrccache

Full path to the running Julia's cache of source code defining Base.

Internal state management

Revise.watched_files — Constant.

Revise.watched_files

Global variable, watched_files[dirname] returns the collection of files in dirname that we're monitoring for changes. The returned value has type WatchList.

This variable allows us to watch directories rather than files, reducing the burden on the OS.

Revise.revision_queue — Constant.

Revise.revision_queue

Global variable, revision_queue holds the names of files that we need to revise, meaning that these files have changed since we last processed a revision. This list gets populated by callbacks that watch directories for updates.

Revise.included_files — Constant.

Revise.included_files

Global variable, included_files gets populated by callbacks we register with include. It's used to track non-precompiled packages and, optionally, user scripts (see docs on JULIA_REVISE_INCLUDE).

Revise.file2modules — Constant.

Revise.file2modules

Global variable, file2modules is the core information that allows re-evaluation of code in the proper module scope. It is a dictionary indexed by absolute paths of files; file2modules[filename] returns a value of type Revise.FileModules.

Revise.module2files — Constant.

Revise.module2files

Global variable, module2files holds the list of filenames used to define a particular module. This is only used by revise(MyModule) to "refresh" all the definitions in a module.

Types

Revise.RelocatableExpr — Type.

A RelocatableExpr is exactly like an Expr except that comparisons between RelocatableExprs ignore line numbering information.

You can use convert(Expr, rex::RelocatableExpr) to convert to an Expr and convert(RelocatableExpr, ex::Expr) for the converse. Beware that the latter operates in-place and is intended only for internal use.

Revise.ExprsSigs — Type.

Revise.ExprsSigs

struct holding parsed source code.

Fields:

exprs: all RelocatableExpr in the module or file

sigs: all detected function signatures (used in method deletion)

These fields are stored as sets so that one can efficiently find the differences between two versions of the same module or file.

Revise.ModDict — Type.

A ModDict is an alias for Dict{Module,ExprsSigs}. It is used to organize expressions according to their module of definition.

See also FileModules.

Revise.FileModules — Type.

FileModules(topmod::Module, md::ModDict, [cachefile::String])

Structure to hold the per-module expressions found when parsing a single file. topmod is the current module when the file is parsed. md holds the evaluatable statements, organized by the module of their occurrence. In particular, if the file defines one or more new modules, then md contains key/value pairs for each module. If the file does not define any new modules, topmod is the only key in md.

Example:

Suppose MyPkg.jl has a file that looks like this:

__precompile__(true)

module MyPkg

foo(x) = x^2

end

Then if this module is loaded from Main, schematically the corresponding fm::FileModules looks something like

fm.topmod = Main
fm.md = Dict(Main=>ExprsSigs(OrderedSet([:(__precompile__(true))]), OrderedSet()),
             Main.MyPkg=>ExprsSigs(OrderedSet([:(foo(x) = x^2)]), OrderedSet([:(foo(x))]))

because the precompile statement occurs in Main, and the definition of foo occurs in Main.MyPkg.

Source cache files

Optionally, a FileModule can also record the path to a cache file holding the original source code. This is applicable only for precompiled modules and Base. (This cache file is distinct from the original source file that might be edited by the developer, and it will always hold the state of the code when the package was precompiled or Julia's Base was built.) For such modules, the ExprsSigs will be empty for any file that has not yet been edited: the original source code gets parsed only when a revision needs to be made.

Source cache files greatly reduce the overhead of using Revise.

To create a FileModules from a source file, see parse_source.

Revise.WatchList — Type.

Revise.WatchList

A struct for holding files that live inside a directory. Some platforms (OSX) have trouble watching too many files. So we watch parent directories, and keep track of which files in them should be tracked.

Fields:

timestamp: mtime of last update

trackedfiles: Set of filenames

Function reference

Functions called during initialization of Revise

Revise.async_steal_repl_backend — Function.

Revise.async_steal_repl_backend()

Wait for the REPL to complete its initialization, and then call steal_repl_backend. This is necessary because code registered with atreplinit runs before the REPL is initialized, and there is no corresponding way to register code to run after it is complete.

Revise.steal_repl_backend — Function.

steal_repl_backend(backend = Base.active_repl_backend)

Replace the REPL's normal backend with one that calls revise before executing any REPL input.

Functions called when you load a new package

Revise.watch_package — Function.

watch_package(id::Base.PkgId)

Start watching a package for changes to the files that define it. This function gets called via a callback registered with Base.require, at the completion of module-loading by using or import.

Revise.parse_pkg_files — Function.

parse_pkg_files(modsym)

This function gets called by watch_package and runs when a package is first loaded. Its job is to organize the files and expressions defining the module so that later we can detect and process revisions.

Revise.init_watching — Function.

Revise.init_watching(files)

For every filename in files, monitor the filesystem for updates. When the file is updated, either revise_dir_queued or revise_file_queued will be called.

Monitoring for changes

These functions get called on each directory or file that you monitor for revisions. These block execution until the file(s) are updated, so you should only call them from within an @async block. They work recursively: once an update has been detected and execution resumes, they schedule a revision (see Revise.revision_queue) and then call themselves on the same directory or file to wait for the next set of changes.

Revise.revise_dir_queued — Function.

revise_dir_queued(dirname)

Wait for one or more of the files registered in Revise.watched_files[dirname] to be modified, and then queue the corresponding files on Revise.revision_queue. This is generally called within an @async.

Revise.revise_file_queued — Function.

revise_file_queued(filename)

Wait for modifications to filename, and then queue the corresponding files on Revise.revision_queue. This is generally called within an @async.

This is used only on platforms (like BSD) which cannot use revise_dir_queued.

Evaluating changes (revising) and computing diffs

Revise.revise_file_now — Function.

Revise.revise_file_now(file)

Process revisions to file. This parses file and computes an expression-level diff between the current state of the file and its most recently evaluated state. It then deletes any removed methods and re-evaluates any changed expressions.

file must be a key in Revise.file2modules

Revise.eval_revised — Function.

succeeded = eval_revised(revmd::ModDict, delete_methods=true)

Evaluate the changes listed in revmd, which consists of deleting all the listed signatures in each .sigs field(s) (unless delete_methods=false) and evaluating expressions in the .exprs field(s).

Returns true if all revisions in revmd were successfully implemented.

Revise.revised_statements — Function.

revmod = revised_statements(new_defs, old_defs)

Return a Dict(Module=>changeset), revmod, listing the changes that should be eval_revised for each module to update definitions from old_defs to new_defs. See parse_source to obtain the defs structures.

Revise.get_method — Function.

m = get_method(mod::Module, sig)

Get the method m with signature sig from module mod. This is used to provide the method to Base.delete_method. See also get_signature.

Parsing source code

Revise.parse_source — Function.

parse_source(filename::AbstractString, mod::Module)

Parse the source filename, returning a pair filename => fm::FileModules (see FileModules) containing the information needed to evaluate code in file. mod is the "parent" module for the file (i.e., the one that included the file); if filename defines more module(s) then these will all have separate entries in fm.md.

If parsing filename fails, nothing is returned.

Revise.parse_source! — Function.

success = parse_source!(md::ModDict, filename, mod::Module)

Top-level parsing of filename as included into module mod. Successfully-parsed expressions will be added to md. Returns true if parsing finished successfully.

See also parse_source.

success = parse_source!(md::ModDict, src::AbstractString, file::Symbol, pos::Integer, mod::Module)

Parse a string src obtained by reading file as a single string. pos is the 1-based byte offset from which to begin parsing src.

See also parse_source.

success = parse_source!(md::ModDict, ex::Expr, file, mod::Module)

For a file that defines a sub-module, parse the body ex of the sub-module. mod will be the module into which this sub-module is evaluated (i.e., included). Successfully-parsed expressions will be added to md. Returns true if parsing finished successfully.

See also parse_source.

Revise.parse_expr! — Function.

parse_expr!(md::ModDict, ex::Expr, file::Symbol, mod::Module)

Recursively parse the expressions in ex, iterating over blocks and sub-module definitions. Successfully parsed expressions are added to md with key mod, and any sub-modules will be stored in md using appropriate new keys. This accomplishes two main tasks:

add parsed expressions to the source-code cache (so that later we can detect changes)

determine the module into which each parsed expression is evaluated into

Revise.parse_module! — Function.

newmod = parse_module!(md::ModDict, ex::Expr, file, mod::Module)

Parse an expression ex that defines a new module newmod. This module is "parented" by mod. Source-code expressions are added to md under the appropriate module name.

Revise.get_signature — Function.

sig = get_signature(expr)

Extract the signature from an expression expr that defines a function.

If expr does not define a function, returns nothing.

Revise.sig_type_exprs — Function.

typexs = sig_type_exprs(ex::Expr)

From a function signature ex (see get_signature), generate a list typexs of concrete signature type expressions. This list will have length 1 unless ex has default arguments, in which case it will produce one type signature per valid number of supplied arguments.

These type-expressions can be evaluated in the appropriate module to obtain a Tuple-type.

Examples

julia> Revise.sig_type_exprs(:(foo(x::Int, y::String)))
1-element Array{Expr,1}:
:(Tuple{Core.Typeof(foo), Int, String})

julia> Revise.sig_type_exprs(:(foo(x::Int, y::String="hello")))
2-element Array{Expr,1}:
 :(Tuple{Core.Typeof(foo), Int})
 :(Tuple{Core.Typeof(foo), Int, String})

Git integration

Revise.git_source — Function.

Revise.git_source(file::AbstractString, reference)

Read the source-text for file from a git commit reference. The reference may be a string, Symbol, or LibGit2.Tree.

Example:

Revise.git_source("/path/to/myfile.jl", "HEAD")
Revise.git_source("/path/to/myfile.jl", :abcd1234)  # by commit SHA

Revise.git_files — Function.

files = git_files(repo)

Return the list of files checked into repo.

Revise.git_repo — Function.

repo, repo_path = git_repo(path::AbstractString)

Return the repo::LibGit2.GitRepo containing the file or directory path. path does not necessarily need to be the top-level directory of the repository. Also returns the repo_path of the top-level directory for the repository.