User reference

User reference

There are really only three functions that a user would be expected to call manually: revise, includet, and Revise.track. Other user-level constructs might apply if you want to exclude Revise from watching specific packages.

Revise.revise — Function.

revise()

eval any changes in the revision queue. See Revise.revision_queue.

source

revise(mod::Module)

Reevaluate every definition in mod, whether it was changed or not. This is useful to propagate an updated macro definition, or to force recompiling generated functions.

source

Revise.track — Function.

Revise.track(Base)
Revise.track(Core.Compiler)
Revise.track(stdlib)

Track updates to the code in Julia's base directory, base/compiler, or one of its standard libraries.

source

Revise.track(mod::Module, file::AbstractString)
Revise.track(file::AbstractString)

Watch file for updates and revise loaded code with any changes. mod is the module into which file is evaluated; if omitted, it defaults to Main.

source

Revise.includet — Function.

includet(filename)

Load filename and track any future changes to it. includet is deliberately non-recursive, so if filename loads any other files, they will not be automatically tracked. (See Revise.track to set it up manually.)

includet is intended for "user scripts," e.g., a file you use locally for a specific purpose such as loading a specific data set or performing some kind of analysis. Do not use includet for packages, as those should be handled by using or import. If using and import aren't working, you may have packages in a non-standard location; try fixing it with something like push!(LOAD_PATH, "/path/to/my/private/repos").

source

Revise.dont_watch_pkgs — Constant.

Revise.dont_watch_pkgs

Global variable, use push!(Revise.dont_watch_pkgs, :MyPackage) to prevent Revise from tracking changes to MyPackage. You can do this from the REPL or from your .julia/config/startup.jl file.

See also Revise.silence.

source

Revise.silence — Function.

Revise.silence(pkg)

Silence warnings about not tracking changes to package pkg.

source